Avoid loading packages in Python tests
