+++ src/Python/qsharp-core/qsharp/tests/test_azure.py
@@ -9,22 +9,16 @@
 
 ## IMPORTS ##
 
-import importlib
-import os
 import pytest
 import qsharp
 from qsharp.azure import AzureError, AzureJob, AzureTarget
-import sys
+from .utils import set_environment_variables
 
 ## SETUP ##
 
 @pytest.fixture(scope="session", autouse=True)
-def set_environment_variables():
-    # Need to restart the IQ# kernel after setting the environment variable
-    os.environ["AZURE_QUANTUM_ENV"] = "mock"
-    importlib.reload(qsharp)
-    if "qsharp.chemistry" in sys.modules:
-        importlib.reload(qsharp.chemistry)
+def session_setup():
+    set_environment_variables()
 
 ## TESTS ##
